refactor: correct platform and system layer ownership
