+++ CMakeLists.txt
@@ -18,8 +18,8 @@ target_sources(pslab_pico PRIVATE
         src/application/protocol/dso.c
         src/application/logic_analyser_commands.c
         src/application/dso_commands.c
-        src/platform/dso.c
-        src/system/adc_capture.c
+        src/system/dso.c
+        src/platform/adc_capture.c
         src/platform/usb_cdc.c
         src/platform/usb_descriptors.c
         src/platform/status_led.c
