+++ src/platform/adc_capture.h
